requirements.txt
wordcloud
matplotlib
sentence_transformers
umap-learn
sklearn
plotly
